# docs/openapi/gateway.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "gateway.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/tasks/end": {
      "post": {
        "operationId": "GatewayService_EndTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gatewayEndTaskResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/gatewayEndTaskRequest"
            }
          }
        ],
        "tags": [
          "GatewayService"
        ]
      }
    },
...
    "/usage/endpoints": {
      "get": {
        "operationId": "GatewayService_GetEndpointUsage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gatewayGetEndpointUsageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "kind",
            "description": "spend (default) | earned",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "description": "Inclusive UTC day bounds (YYYY-MM-DD); end_date defaults to today.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GatewayService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "gatewayEndTaskRequest": {
      "type": "object",
      "properties": {
        "taskId": {
          "type": "string"
        },
        "taskDuration": {
          "type": "number",
          "format": "float"
        },
        "taskStatus": {
          "type": "string"
        },
        "containerId": {
          "type": "string"
        },
        "containerHostname": {
          "type": "string"
        },
        "keepWarmSeconds": {
          "type": "number",
          "format": "float"
        },
        "result": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "gatewayEndTaskResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean"
        }
      }
    },
    "gatewayEndpointUsage": {
      "type": "object",
      "properties": {
        "requests": {
          "type": "string",
          "format": "int64"
        },
        "promptTokens": {
          "type": "string",
          "format": "int64"
        },
        "completionTokens": {
          "type": "string",
          "format": "int64"
        },
        "images": {
          "type": "string",
          "format": "int64"
        },
        "microUsd": {
          "type": "string",
          "format": "int64"
        },
        "cachedTokens": {
          "type": "string",
          "format": "int64",
          "description": "Cached tokens are included in prompt_tokens. Prompt cost excludes cache."
        },
        "promptMicroUsd": {
          "type": "string",
          "format": "int64"
        },
        "completionMicroUsd": {
          "type": "string",
          "format": "int64"
        },
        "cachedMicroUsd": {
          "type": "string",
          "format": "int64"
        },
        "requestMicroUsd": {
          "type": "string",
          "format": "int64"
        },
        "imageMicroUsd": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "gatewayGetEndpointUsageResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean"
        },
        "errMsg": {
          "type": "string"
        },
        "total": {
          "$ref": "#/definitions/gatewayEndpointUsage"
        },
        "perModel": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/gatewayEndpointUsage"
          }
        },
        "perDay": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/gatewayEndpointUsage"
          }
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
  }
}